requirements.txt
requests
pandas
SQLAlchemy
python-dotenv
psycopg2-binary